+++ CHANGES.md
@@ -1,5 +1,10 @@
 ### 1.0.0:
 
+(Oct 18, 2018)
+
+- Added basic configuration
+- Verbose flag
+
 (Oct 14, 2018)
 
 - First version